\name{paternityDP-package}
\alias{paternityDP-package}
\alias{paternityDP}
\docType{package}
\title{
\packageTitle{paternityDP}
}
\description{
\packageDescription{paternityDP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{paternityDP}
\packageIndices{paternityDP}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{paternityDP}

Maintainer: \packageMaintainer{paternityDP}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
