\name{sydneyPaternity-package}
\alias{sydneyPaternity-package}
\alias{sydneyPaternity}
\docType{package}
\title{
\packageTitle{sydneyPaternity}
}
\description{
\packageDescription{sydneyPaternity}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sydneyPaternity}
\packageIndices{sydneyPaternity}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{sydneyPaternity}

Maintainer: \packageMaintainer{sydneyPaternity}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
